Break out utils to separate module
